D3D12 ShaderFeedback fix GPU hang when analysing ExecuteIndrect
Need to recreate the EI command signature using the same root signature that is is bound i.e. the modified root signature which includes the resources for the shader feedback analysis

This only needs to be done if the EI command signature modifies the root arguments i.e. setting root constants, updating bindings